CV/project-2/requirements.txt
ipdb==0.8
ipython==3.1.0
matplotlib==1.4.3
nose==1.3.6
numpy==1.9.2
Pillow==2.8.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
scipy==0.15.1
six==1.9.0
